Check for JID user part

+++ kontalk/xmppserver/component/c2s.py
@@ -397,7 +397,7 @@ class InitialPresenceHandler(XMPPHandler):
         self.parent.sfactory.check_conflict(sender)
 
         # initial presence from a client connected to another server, clear it from our presence table
-        if sender.host != self.parent.servername and sender.host in self.parent.keyring.hostlist():
+        if sender.user and sender.host != self.parent.servername and sender.host in self.parent.keyring.hostlist():
             log.debug("deleting %s from presence table" % (sender.user, ))
             self.parent.presencedb.delete(sender.user)
